refactor: split head and provider to file
